requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==4.0.1
Django==4.2.1
django-debug-toolbar==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
Faker==18.9.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
